requirements.txt
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
constantly==15.1.0
coreapi==2.3.1
coreapi-cli==1.0.6
coreschema==0.0.4
cryptography==2.0.3
cssselect==1.0.1
Django==1.11.4
django-codemirror-widget==0.4.1
django-crispy-forms==1.6.1
django-filter==1.0.4
django-pagedown==0.1.3
djangorestframework==3.6.3
httpie==0.9.9
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.9.6
lxml==3.8.0
Markdown==2.6.8
markdown2==2.3.4
MarkupSafe==1.0
mysqlclient==1.3.10
parsel==1.2.0
pyasn1==0.3.2
pyasn1-modules==0.0.11
pycparser==2.18
PyDispatcher==2.0.5
Pygments==2.2.0
pyOpenSSL==17.2.0
pytz==2017.2
queuelib==1.4.2
requests==2.18.4
Scrapy==1.4.0
service-identity==17.0.0
six==1.10.0
Twisted==17.5.0
uritemplate==3.0.0
urllib3==1.22
w3lib==1.18.0
zope.interface==4.4.2
